Remove Mint-X and Linux Mint themes
Removed Mint-X and Linux Mint themes bacause they look old-school
and no one would want to use them.

+++ generate-themes.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python3
 import os
 
-from constants import X_HEX_ACCENTS, X_RGB_ACCENTS, x_hex_colors, x_rgb_colors
 from constants import Y_HEX_ACCENT1, Y_HEX_ACCENT2, Y_HEX_ACCENT3, Y_HEX_ACCENT4
 from constants import y_hex_colors1, y_hex_colors2, y_hex_colors3, y_hex_colors4
 
@@ -11,12 +10,6 @@ def change_value (key, value, file):
     else:
         command = "sed -i '/%(key)s=/d' %(file)s" % {'key':key, 'file':file}
     os.system(command)
-
-def x_colorize_directory (path, variation):
-    for accent in X_HEX_ACCENTS:
-        os.system("find %s -name '*.*' -type f -exec sed -i 's/%s/%s/gI' {}  \\;" % (path, accent, x_hex_colors[variation]))
-    for accent in X_RGB_ACCENTS:
-        os.system("find %s -name '*.*' -type f -exec sed -i 's/%s/%s/gI' {}  \\;" % (path, accent, x_rgb_colors[variation]))
 
 def y_colorize_directory (path, variation):
     for accent in Y_HEX_ACCENT1:
@@ -33,91 +26,69 @@ if os.path.exists("usr"):
 
 os.system("mkdir -p usr/share/themes")
 
-os.system("cp -R src/Mint-X/theme/* usr/share/themes/")
-
-for color in os.listdir("src/Mint-X/variations"):
-    path = os.path.join("src/Mint-X/variations", color)
-    if os.path.isdir(path):
-        theme = "usr/share/themes/Mint-X-%s" % color
-        theme_index = os.path.join(theme, "index.theme")
-        os.system("cp -R usr/share/themes/Mint-X %s" % theme)
-        os.system("cp -R src/Mint-X/variations/%s/* %s/" % (color, theme))
-
-        # Theme name
-        for key in ["Name", "GtkTheme", "IconTheme"]:
-            change_value(key, "Mint-X-%s" % color, theme_index)
-
-        # Accent color
-        gtkrc = os.path.join(theme, "gtk-2.0", "gtkrc")
-        settings_ini = os.path.join(theme, "gtk-3.0", "settings.ini")
-        gtk_main_css = os.path.join(theme, "gtk-3.0", "gtk-main.css")
-        for file in [gtkrc, settings_ini, gtk_main_css]:
-            for accent in X_HEX_ACCENTS:
-                os.system("sed -i s'/%(accent)s/%(color_accent)s/' %(file)s" % {'accent': accent, 'color_accent': x_hex_colors[color], 'file': file})
-
-        # Cinnamon theme name
-        file = os.path.join(theme, "cinnamon", "theme.json")
-        if os.path.exists(file):
-            os.system("sed -i s'/Mint-X/Mint-X-%(color)s/' %(file)s" % {'color': color, 'file': file})
-
-        # Cinnamon colors
-        file = os.path.join(theme, "cinnamon", "cinnamon.css")
-        if os.path.exists(file):
-            for accent in X_HEX_ACCENTS:
-                os.system("sed -i s'/%(accent)s/%(color_accent)s/' %(file)s" % {'accent': accent, 'color_accent': x_hex_colors[color], 'file': file})
-            for accent in X_RGB_ACCENTS:
-                os.system("sed -i s'/%(accent)s/%(color_accent)s/' %(file)s" % {'accent': accent, 'color_accent': x_rgb_colors[color], 'file': file})
-
 curdir = os.getcwd()
 
+# Build Mint-Y Base themes (light, dark, darker)
 os.chdir("src/Mint-Y")
 os.system("./build-themes.py")
+
 os.chdir(curdir)
 
 # Mint-Y color variations
 for color in y_hex_colors1.keys():
-    for variant in ["", "-Dark"]:
+    for variant in ["", "-Dark", "-Darker"]:
         original_name = "Mint-Y%s" % variant
         path = os.path.join("src/Mint-Y/variations/%s" % color)
         if os.path.isdir(path):
             print("Derivating %s-%s" % (original_name, color))
-
+            
             # Copy theme
             theme = "usr/share/themes/%s-%s" % (original_name, color)
             theme_index = os.path.join(theme, "index.theme")
             os.system("cp -R usr/share/themes/%s %s" % (original_name, theme))
-
+            
             # Theme name
             for key in ["Name", "GtkTheme"]:
                 change_value(key, "%s-%s" % (original_name, color), theme_index)
-
+            
+            for key in ["MetacityTheme"]:
+                metacity_variant = original_name.replace("Darker", "Dark")
+                change_value(key, "%s-%s" % (metacity_variant, color), theme_index)
+            
             for key in ["IconTheme"]:
                 change_value(key, "%s-%s" % (original_name, color), theme_index)
-
+            
+            for key in ["CursorTheme"]:
+                change_value(key, "FreedomOS-%s" % color, theme_index)
+            
             # Regenerate GTK3 sass
             os.system("cp -R src/Mint-Y/gtk-3.0/sass %s/gtk-3.0/" % theme)
             y_colorize_directory("%s/gtk-3.0/sass" % theme, color)
             os.chdir("%s/gtk-3.0" % theme)
             if (variant == "-Dark"):
                 os.system("cp sass/gtk-dark.scss sass/gtk.scss")
-                os.system("sassc ./sass/gtk.scss gtk.css")
+            elif (variant == "-Darker"):
+                os.system("cp sass/gtk-darker.scss sass/gtk.scss")
             else:
-                os.system("sassc ./sass/gtk-dark.scss gtk-dark.css")
-                os.system("sassc ./sass/gtk.scss gtk.css")
-
+                os.system("rm sass/gtk-dark.scss sass/gtk-darker.scss")
+            
+            os.system("sassc ./sass/gtk.scss gtk.css")
             os.system("rm -rf sass .sass-cache")
             os.chdir(curdir)
-
+            
             # Regenerate Cinnamon sass
-            os.system("cp -R src/Mint-Y/cinnamon/sass %s/cinnamon/" % theme)
-            y_colorize_directory("%s/cinnamon/sass" % theme, color)
-            os.chdir("%s/cinnamon" % theme)
-            if (variant == "-Dark"):
-                os.system("cp sass/cinnamon-dark.scss sass/cinnamon.scss")
-            os.system("sassc ./sass/cinnamon.scss cinnamon.css")
-            os.system("rm -rf sass .sass-cache")
+            if (variant != "-Darker"):
+                # Darker variants have no cinnamon style
+                os.system("cp -R src/Mint-Y/cinnamon/sass %s/cinnamon/" % theme)
+                os.system("cp -R %s/cinnamon/sass %s/cinnamon/" % (path, theme))
+                y_colorize_directory("%s/cinnamon/sass" % theme, color)
+                os.chdir("%s/cinnamon" % theme)
+                if (variant == "-Dark"):
+                    os.system("cp sass/cinnamon-dark.scss sass/cinnamon.scss")
+                os.system("sassc ./sass/cinnamon.scss cinnamon.css")
+                os.system("rm -rf sass .sass-cache")
             os.chdir(curdir)
-
+            
             # Accent color
             files = []
             files.append(os.path.join(theme, "gtk-2.0", "gtkrc"))
@@ -135,10 +106,10 @@ for color in y_hex_colors1.keys():
                         os.system("sed -i s'/%(accent)s/%(color_accent)s/gI' %(file)s" % {'accent': accent, 'color_accent': y_hex_colors3[color], 'file': file})
                     for accent in Y_HEX_ACCENT4:
                         os.system("sed -i s'/%(accent)s/%(color_accent)s/gI' %(file)s" % {'accent': accent, 'color_accent': y_hex_colors4[color], 'file': file})
-
-            # Remove metacity-theme-3.xml (it doesn't need to be derived since it's using GTK colors, and Cinnamon doesn't want to list it)
-            os.system("rm -f %s" % os.path.join(theme, "metacity-1", "metacity-theme-3.xml"))
-
+            
+            # # Remove metacity-theme-3.xml (it doesn't need to be derived since it's using GTK colors, and Cinnamon doesn't want to list it)
+            # os.system("rm -f %s" % os.path.join(theme, "metacity-1", "metacity-theme-3.xml"))
+            
             directories = []
             directories.append(os.path.join(theme, "cinnamon/common-assets"))
             directories.append(os.path.join(theme, "cinnamon/light-assets"))
@@ -146,7 +117,7 @@ for color in y_hex_colors1.keys():
             for directory in directories:
                 if os.path.exists(directory):
                     y_colorize_directory(directory, color)
-
+            
             # Assets
             os.system("rm -rf %s/gtk-3.0/assets" % theme)
             os.system("rm -rf %s/gtk-2.0/assets" % theme)
@@ -155,7 +126,6 @@ for color in y_hex_colors1.keys():
             else:
                 os.system("cp -R %s/gtk-2.0/assets %s/gtk-2.0/assets" % (path, theme))
             os.system("cp -R %s/gtk-3.0/assets %s/gtk-3.0/assets" % (path, theme))
+            
+            os.system("cp -R files/%s ./usr/share/themes" % theme)
 
-
-# Files
-os.system("cp -R files/* ./")
